+++ R/stat-bin.r
@@ -121,12 +121,16 @@ StatBin <- ggproto("StatBin", Stat,
   compute_group = function(data, scales, binwidth = NULL, bins = NULL,
                            center = NULL, boundary = NULL,
                            closed = c("right", "left"), pad = FALSE,
+                           breaks = NULL,
                            # The following arguments are not used, but must
                            # be listed so parameters are computed correctly
-                           breaks = NULL, origin = NULL, right = NULL,
-                           drop = NULL, width = NULL) {
+                           origin = NULL, right = NULL, drop = NULL,
+                           width = NULL) {
 
     if (!is.null(breaks)) {
+      if (!scales$x$is_discrete()){
+         breaks <- scales$x$transform(breaks)
+      }
       bins <- bin_breaks(breaks, closed)
     } else if (!is.null(binwidth)) {
       if (is.function(binwidth)) {